backend/pyproject.toml
[tool.poetry]
name = "cr-backend"
version = "0.1.0"
description = "Retail POS backend"
authors = ["OpenAI"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras=["standard"], version="^0.30.1"}
sqlalchemy = "^2.0.30"
asyncpg = "^0.29.0"
psycopg = {extras=["binary"], version="^3.1.19"}
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
authlib = "^1.3.0"
passlib = {extras=["bcrypt"], version="^1.7.4"}
python-jose = "^3.3.0"
aiofiles = "^23.2.1"
httpx = "^0.27.0"
python-multipart = "^0.0.9"
alembic = "^1.13.2"
aiosqlite = "^0.20.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ruff = "^0.4.8"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100

[tool.black]
line-length = 100
